Add support for sub-templates.
Implementation of issue #86. Sub-templates are just like normal project
templates, except they are not published separately. Instead, they are included
directly in a project template package.

The user can run the new `generate` command to activate a sub-template whose
job it is to actually do something, such as create new files in the project.
The user can also pass in parameters if the sub-template uses the `ask()`
method.

The Gradle plugin for managing Lazybones templates now allows you to define
relationships between main templates and sub-templates. The sub-templates are
packaged separately and then included into the main template package zip.

You create sub-templates as normal templates except that the directory name
must begin with 'subtmpl-'. This ensures that they aren't included in the
`packageAllTemplates`, `installAllTemplates` and `publishAllTemplates` as they
shouldn't be published independently. You then use the Lazybones extension to
declare that a template includes some set of sub-templates. The plugin does
the rest.

+++ lazybones-app/src/integTest/groovy/uk/co/cacoethes/lazybones/AbstractFunctionalSpec.groovy
@@ -40,9 +40,11 @@ abstract class AbstractFunctionalSpec extends Specification {
      * larger.
      * @return The exit code of the lazybones process.
      */
-    int runCommand(List cmdList, File workDir, List inputs = []) {
-        // Clear out results of previous executions
-        FileUtils.deleteDirectory(workDir)
+    int runCommand(List cmdList, File workDir, List inputs = [], boolean clearPrevious = true) {
+        if (clearPrevious) {
+            // Clear out results of previous executions
+            FileUtils.deleteDirectory(workDir)
+        }
         resetOutput()
         workDir.mkdirs()
 
@@ -72,7 +74,7 @@ abstract class AbstractFunctionalSpec extends Specification {
         // the form VAR=value.
         def envp = env.collect { key, value -> key + "=" + value }
 
-        Process process = ([lzbExecutable, "--stacktrace"] + cmdList).execute(envp, workDir)
+        Process process = ([lzbExecutable] + cmdList).execute(envp, workDir)
 
         if (inputs) {
             def newLine = System.getProperty("line.separator")
